requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
drf_yasg
psycopg2
python-dotenv
djangorestframework-simplejwt
django-cors-headers
PyJWT